refactor: group admin workflow adapters
