+++ crates/dcc-mcp-gateway/src/gateway/admin/router.rs
@@ -9,6 +9,10 @@ use super::analytics::{
 use super::application::agent_trace::handle_v1_debug_agent_trace_packet;
 use super::application::artifacts::handle_admin_artifacts;
 use super::application::events::handle_admin_events;
+use super::application::experiments::{
+    handle_experiment_create, handle_experiment_detail, handle_experiment_judge_result,
+    handle_experiment_run, handle_experiments_list,
+};
 use super::application::general::{
     handle_admin_activity, handle_admin_governance, handle_admin_traffic,
     handle_admin_traffic_export, handle_admin_ui,
@@ -22,26 +26,22 @@ use super::application::handlers::{
     handle_v1_debug_trace_lookup,
 };
 use super::application::health::{handle_admin_health, handle_admin_reliability};
-use super::application::memory::{handle_admin_memory, handle_admin_memory_forget};
-use super::application::sessions::{handle_admin_session_detail, handle_admin_sessions};
-use super::application::skill_paths::{
-    handle_admin_skill_path_add, handle_admin_skill_path_delete, handle_admin_skill_paths,
-};
-use super::experiments::{
-    handle_experiment_create, handle_experiment_detail, handle_experiment_judge_result,
-    handle_experiment_run, handle_experiments_list,
-};
-use super::infra::integrations::{
-    handle_admin_integration_test, handle_admin_integration_update, handle_admin_integrations,
-};
-use super::marketplace::{
+use super::application::marketplace::{
     handle_marketplace_add_source, handle_marketplace_catalog, handle_marketplace_install,
     handle_marketplace_installed, handle_marketplace_outdated, handle_marketplace_sources,
     handle_marketplace_uninstall, handle_marketplace_update,
 };
-use super::recordings::{
+use super::application::memory::{handle_admin_memory, handle_admin_memory_forget};
+use super::application::recordings::{
     handle_recording_compile, handle_recording_replay_validate, handle_recording_review,
     handle_recording_review_body, handle_recording_start, handle_recording_stop,
+};
+use super::application::sessions::{handle_admin_session_detail, handle_admin_sessions};
+use super::application::skill_paths::{
+    handle_admin_skill_path_add, handle_admin_skill_path_delete, handle_admin_skill_paths,
+};
+use super::infra::integrations::{
+    handle_admin_integration_test, handle_admin_integration_update, handle_admin_integrations,
 };
 use super::state::AdminState;
